requirements.txt
# Core Django dependencies
Django==5.0.3
djangorestframework==3.15.2

# Database (MySQL)
mysqlclient==2.2.4

# For making HTTP requests to external APIs
requests==2.32.3

# For image generation (summary.png)
Pillow==10.3.0

# For background tasks or scheduling (optional but useful)
celery==5.4.0

# For CORS (if you test with frontend later)
django-cors-headers==4.4.0

# Optional utilities for better debugging and performance
django-extensions==3.2.3
drf-yasg==1.21.7  # For API documentation (Swagger)
